add: server config for production
