+++ .env.example
@@ -1,4 +1,6 @@
 VITE_API_URL=https://acc-jungle-gym-api.herokuapp.com
 VITE_STRAPI_API_URL=https://jungle-gym-cms.herokuapp.com
 VITE_SESSION_DB_URL=localhost
+VITE_SESSION_DB_USERNAME=false
 VITE_SESSION_DB_PORT=7001
+VITE_SESSION_DB_PASSWORD=false
