+++ README.md
@@ -33,6 +33,8 @@ The **Meesterproef** is the final client-project of the half year minor programm
 
 ### Tech stack
 
+## Prerequisites
+* Redis database for session storage
 ## :gear: Installation
 
 ### Developing
