+++ src/hooks.ts
@@ -1,25 +1,5 @@
-import { SESSION_DB_PORT, SESSION_DB_URL } from "$lib/Env";
-import { Tedis} from 'tedis'
 import * as cookie from 'cookie'
-import type { GetSession, Handle } from "@sveltejs/kit";
-
-const host: string | boolean = SESSION_DB_URL
-const port: number | boolean = Number(SESSION_DB_PORT)
-
-
-let sessionDB: Tedis | undefined
-let db_error;
-
-
-if(typeof port === 'number' && !isNaN(port) && typeof host === 'string' ) {
-	sessionDB = new Tedis({host, port})
-
-	sessionDB.on('connect', () => console.log('SessionDB connected'))
-	sessionDB.on('error', (error) => {
-		db_error = error
-	})
-	sessionDB.on('close', (had_error) => console.log('SessionDB closed', had_error))
-}
+import sessionDB from './routes/account/_session'
 
 export const handle: Handle = async ({request, render}) => {
 	const cookies = cookie.parse(request.headers.cookie || '')
@@ -38,7 +18,7 @@ export const handle: Handle = async ({request, render}) => {
 export const getSession: GetSession = async (request) => {
 	if(!sessionDB) return {
 		user: null,
-		error: db_error,
+		error: 'SessionDB offline',
 		authenticated: false
 	}
