+++ src/lib/Env.js
@@ -2,3 +2,5 @@
 export const BASE_API_URL = import.meta.env.VITE_STRAPI_API_URL;
 export const SESSION_DB_URL = import.meta.env.VITE_SESSION_DB_URL;
 export const SESSION_DB_PORT = import.meta.env.VITE_SESSION_DB_PORT;
+export const SESSION_DB_USERNAME = import.meta.env.VITE_SESSION_DB_USERNAME;
+export const SESSION_DB_PASSWORD = import.meta.env.VITE_SESSION_DB_PASSWORD;
