+++ src/routes/account/inloggen.json.ts
@@ -1,28 +1,9 @@
 import type { Request, RequestHandler } from "@sveltejs/kit";
 import { user } from "./_mockAccounts";
-import { Tedis} from 'tedis'
 import { hash } from 'bcrypt'
 import { v4 as uuidv4 } from 'uuid'
 import * as cookie from 'cookie'
-import { SESSION_DB_PORT, SESSION_DB_URL } from "$lib/Env";
-
-const host: string | boolean = SESSION_DB_URL
-const port: number | boolean = Number(SESSION_DB_PORT)
-
-
-let sessionDB: Tedis | undefined
-let db_error;
-
-
-if(typeof port === 'number' && !isNaN(port) && typeof host === 'string' ) {
-	sessionDB = new Tedis({host, port})
-
-	sessionDB.on('connect', () => console.log('SessionDB connected'))
-	sessionDB.on('error', (error) => {
-		db_error = error
-	})
-	sessionDB.on('close', (had_error) => console.log('SessionDB closed', had_error))
-}
+import sessionDB from './_session'
 
 
 export const post: RequestHandler = async (request: Request) => {
@@ -31,7 +12,7 @@ export const post: RequestHandler = async (request: Request) => {
 	
 	if(!sessionDB) return {
 		status: 500,
-		body: { message: 'Database Error', db_error}
+		body: { message: 'SessionDB offline'}
 	}
 
 	if(body.username !== user.username || body.password !== user.password) return {
